refactor: split gauge calculoator files
