+++ src/main/kotlin/com/kroffle/knitting/infra/design/calculator/data/GaugeType.kt
@@ -1,0 +1,6 @@
+package com.kroffle.knitting.infra.design.calculator.data
+
+enum class GaugeType {
+    Stitch,
+    Row
+}
