+++ src/main/kotlin/com/kroffle/knitting/infra/design/calculator/data/Method.kt
@@ -1,0 +1,7 @@
+package com.kroffle.knitting.infra.design.calculator.data
+
+enum class Method {
+    Round,
+    RoundUp,
+    RoundDown
+}
