+++ src/test/kotlin/com/kroffle/knitting/infra/calculator/PatternCalculatorTest.kt
@@ -1,10 +1,10 @@
-package com.kroffle.knitting.infra
+package com.kroffle.knitting.infra.calculator
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 import com.kroffle.knitting.domain.design.value.Gauge
 import com.kroffle.knitting.domain.design.value.Pattern
-import com.kroffle.knitting.infra.design.PatternCalculator
+import com.kroffle.knitting.infra.design.calculator.PatternCalculator
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
